+++ index.js
@@ -29,16 +29,17 @@ const ValidationTypes = require('./schemas/validation') // avro related validati
 
 /**
  * @typedef {Object} Module
- * @property {object} rawJSON - Module's dat.json object
+ * @property {object} rawJSON - Module's index.json object
  * @property {object} metadata - Module's metadata, this can contain latest version and modification time.
  */
 
 const {
   InvalidKeyError,
   ValidationError,
-  MissingParam
+  MissingParam,
+  EBUSYError
 } = require('./lib/errors')
-const { createDatJSON, collect } = require('./lib/utils')
+const { createIndexJSON, collect } = require('./lib/utils')
 
 // helper assert fn
 const assertValid = (type, val) => {
@@ -107,6 +108,9 @@ class SDK {
     this.stores = new Map() // deprecated
     this.drivesToWatch = new Map()
 
+    this.dht = finalOpts.dht
+    this.bootstrap = finalOpts.bootstrap
+
     // start hyperswarm
     this.disableSwarm = !!finalOpts.disableSwarm
     if (!this.disableSwarm) {
@@ -128,6 +132,7 @@ class SDK {
     debug(`dbPath: ${this.dbPath}`)
     debug(`persist drives? ${!!this.persist}`)
     debug(`swarm enabled? ${!this.disableSwarm}`)
+    debug(`watch enabled? ${this.watch}`)
   }
 
   allowedProperties () {
@@ -277,7 +282,7 @@ class SDK {
 
   async getOptionsOrCreate () {
     // read global settings or create with default values according to:
-    // https://github.com/p2pcommons/specs/blob/master/interoperability.md#global-settings
+    // https://github.com/p2pcommons/specs/blob/main/interoperability.md#global-settings
     let options = { ...DEFAULT_GLOBAL_SETTINGS }
     try {
       const optionsFile = await readFile(
@@ -400,7 +405,7 @@ class SDK {
         const moduleDir = join(this.baseDir, urlString)
 
         const statDrive = await drive.stat('/')
-        const statDir = await statFn(join(moduleDir, 'dat.json'))
+        const statDir = await statFn(moduleDir)
         const statDriveMtime = statDrive[0].mtime
         const statDirMtime = statDir.mtime
 
@@ -411,7 +416,8 @@ class SDK {
 
         const driveWatch = await dat.importFiles(drive, moduleDir, {
           watch: this.watch,
-          keepExisting: true
+          keepExisting: false,
+          dereference: true
         })
 
         if (this.watch) {
@@ -419,11 +425,14 @@ class SDK {
             DatEncoding.encode(drive.discoveryKey),
             driveWatch
           )
-
-          // NOTE(dk): consider return the driveWatch EE.
-          // Note(dk): consider add a timeout
-          await once(driveWatch, 'put-end')
         }
+
+        const unwatch = await new Promise(resolve => {
+          const unwatch = drive.watch('', () => {
+            resolve(unwatch)
+          })
+        })
+        unwatch.destroy()
 
         if (metadata.lastModified.getTime() >= mtime.getTime()) continue
 
@@ -434,9 +443,9 @@ class SDK {
           if (dirFiles.length === files.length) continue
         }
 
-        // update dat.json file
-        const module = await drive.readFile('dat.json')
-        const datJSON = this._unflatten(JSON.parse(module))
+        // update index.json file
+        const module = await drive.readFile('index.json')
+        const indexJSON = this._unflatten(JSON.parse(module))
 
         // update metadata
         metadata.lastModified = mtime
@@ -444,8 +453,8 @@ class SDK {
         // update localdb
         await this.localdb.put(urlString, {
           ...metadata,
-          rawJSON: datJSON,
-          avroType: this._getAvroType(datJSON.p2pcommons.type).name
+          rawJSON: indexJSON,
+          avroType: this._getAvroType(indexJSON.p2pcommons.type).name
         })
       } catch (err) {
         this._log(`refreshMTimes: ${err.message}`, 'error')
@@ -460,12 +469,17 @@ class SDK {
       storageOpts: { storageLocation: this.baseDir },
       corestoreOpts: {}
     })
+
+    this.store.on('error', err => {
+      this._log(err.message, 'error')
+    })
   }
 
   async startSwarm () {
     if (!this.disableSwarm) {
       this.networker = this.swarmFn(this.store, {
-        announceLocalAddress: true
+        announceLocalAddress: true,
+        bootstrap: this.bootstrap
       })
       this.networker.on('error', console.error)
       if (typeof this.networker.listen === 'function') {
@@ -545,7 +559,7 @@ class SDK {
     // follow module spec: https://github.com/p2pcommons/specs/pull/1/files?short_path=2d471ef#diff-2d471ef4e3a452b579a3367eb33ccfb9
     // 1. create folder with unique name (pk)
     // 2. initialize an hyperdrive inside
-    // 3. createDatJSON with the correct metadata and save it there
+    // 3. createIndexJSON with the correct metadata and save it there
     //
     assert(typeof type === 'string', ValidationError, 'string', type, 'type')
     assert(
@@ -586,8 +600,8 @@ class SDK {
     await ensureDir(moduleDir)
     debug(`ensure module dir: ${moduleDir}`)
 
-    // create dat.json metadata
-    const datJSON = createDatJSON({
+    // create index.json metadata
+    const indexJSON = createIndexJSON({
       type,
       title,
       subtype,
@@ -604,31 +618,41 @@ class SDK {
     // Note(dk): validate earlier
     const avroType = this._getAvroType(type)
 
-    assertValid(avroType, datJSON)
+    assertValid(avroType, indexJSON)
 
-    // write dat.json
-    await writeFile(join(moduleDir, 'dat.json'), JSON.stringify(datJSON))
+    // write index.json
+    await writeFile(join(moduleDir, 'index.json'), JSON.stringify(indexJSON))
 
     const driveWatch = await dat.importFiles(archive, moduleDir, {
-      watch: this.watch
+      watch: this.watch,
+      keepExisting: false,
+      dereference: true
     })
 
     if (this.watch) {
+      driveWatch.on('error', err => {
+        if (err.code === 'EBUSY') {
+          throw new EBUSYError(err.message, publicKeyString)
+        }
+      })
       this.drivesToWatch.set(
         DatEncoding.encode(archive.discoveryKey),
         driveWatch
       )
-
-      // NOTE(dk): consider return the driveWatch EE.
-      // Note(dk): consider add a timeout
-      await once(driveWatch, 'put-end')
     }
 
+    const unwatch = await new Promise(resolve => {
+      const unwatch = archive.watch('index.json', () => {
+        resolve(unwatch)
+      })
+    })
+    unwatch.destroy()
+
     this._log(
-      `Initialized new ${datJSON.p2pcommons.type}, dat://${publicKeyString}`
+      `Initialized new ${indexJSON.p2pcommons.type}, dat://${publicKeyString}`
     )
 
-    debug('init datJSON', datJSON)
+    debug('init indexJSON', indexJSON)
 
     const stat = await archive.stat('/')
     const metadata = {
@@ -640,21 +664,21 @@ class SDK {
 
     await this.localdb.put(publicKeyString, {
       ...metadata,
-      rawJSON: datJSON,
+      rawJSON: indexJSON,
       avroType: avroType.name
     })
 
     await this._seed(archive)
 
     this._log(
-      `Saved new ${datJSON.p2pcommons.type}, with key: ${publicKeyString}`
+      `Saved new ${indexJSON.p2pcommons.type}, with key: ${publicKeyString}`
     )
     // Note(dk): flatten p2pcommons obj in order to have a more symmetrical API
-    return { rawJSON: this._flatten(datJSON), metadata, driveWatch }
+    return { rawJSON: this._flatten(indexJSON), metadata, driveWatch }
   }
 
-  async saveItem ({ isWritable, lastModified, version, datJSON }) {
-    debug('saveItem', datJSON)
+  async saveItem ({ isWritable, lastModified, version, indexJSON }) {
+    debug('saveItem', indexJSON)
     assert(
       typeof isWritable === 'boolean',
       ValidationError,
@@ -663,15 +687,15 @@ class SDK {
       'isWritable'
     )
     assert(
-      typeof datJSON === 'object',
+      typeof indexJSON === 'object',
       ValidationError,
       'object',
-      datJSON,
-      'datJSON'
+      indexJSON,
+      'indexJSON'
     )
 
-    const keyString = DatEncoding.encode(datJSON.url)
-    const datJSONDir = join(this.baseDir, keyString)
+    const keyString = DatEncoding.encode(indexJSON.url)
+    const indexJSONDir = join(this.baseDir, keyString)
 
     let archive
     debug(`saveItem: looking drive ${keyString} in local structure...`)
@@ -681,7 +705,7 @@ class SDK {
     archive = this.drives.get(version ? `${dkey}+${version}` : dkey)
     if (!archive) {
       debug(`saveItem: calling dat open ${keyString}`)
-      const drive = dat.open(this.store, DatEncoding.decode(datJSON.url), {
+      const drive = dat.open(this.store, DatEncoding.decode(indexJSON.url), {
         sparse: this.globalOptions.sparse,
         sparseMetadata: this.globalOptions.sparseMetadata
       })
@@ -691,18 +715,21 @@ class SDK {
 
     let stat, mtime
     if (isWritable) {
-      await writeFile(join(datJSONDir, 'dat.json'), JSON.stringify(datJSON))
-      await dat.importFiles(archive, datJSONDir)
+      await writeFile(
+        join(indexJSONDir, 'index.json'),
+        JSON.stringify(indexJSON)
+      )
+      await dat.importFiles(archive, indexJSONDir)
 
       version = archive.version
-      stat = await archive.stat('/dat.json')
+      stat = await archive.stat('/index.json')
       mtime = stat[0].mtime
     }
 
     debug('updating cache value')
     // Note(dk): we are almost all the time fetching profiles without indicating any specific version so this is more effective
     this.drives.set(
-      version && datJSON.p2pcommons.type !== 'profile'
+      version && indexJSON.p2pcommons.type !== 'profile'
         ? `${dkey}+${version}`
         : dkey,
       archive
@@ -711,16 +738,16 @@ class SDK {
     const lastM =
       mtime && mtime.getTime() >= lastModified.getTime() ? mtime : lastModified
     debug('saving item on local db')
-    await this.localdb.put(DatEncoding.encode(datJSON.url), {
+    await this.localdb.put(DatEncoding.encode(indexJSON.url), {
       isWritable,
       lastModified: lastM,
       version: version,
-      rawJSON: datJSON,
-      avroType: this._getAvroType(datJSON.p2pcommons.type).name
+      rawJSON: indexJSON,
+      avroType: this._getAvroType(indexJSON.p2pcommons.type).name
     })
 
     return {
-      rawJSON: this._flatten(datJSON),
+      rawJSON: this._flatten(indexJSON),
       metadata: {
         isWritable,
         lastModified: lastM,
@@ -829,7 +856,7 @@ class SDK {
 
     if (params.contents && original.type === 'profile') {
       for (const c of params.contents) {
-        await this.validatePublish(c, original.url)
+        await this.validateRegister(c, original.url)
         isUnique(original.contents, c, 'contents')
       }
     }
@@ -842,7 +869,7 @@ class SDK {
    *
    * @public
    * @async
-   * @link https://github.com/p2pcommons/specs/blob/master/module.md
+   * @link https://github.com/p2pcommons/specs/blob/main/module.md
    * @param {Object} module - Object containing field to update
    * @param {(String|Buffer)} module.url - module dat url REQUIRED
    * @param {String} [module.title]
@@ -935,7 +962,7 @@ class SDK {
     debug('set: valid input')
     return this.saveItem({
       ...metadata,
-      datJSON: finalJSON
+      indexJSON: finalJSON
     })
   }
 
@@ -1199,24 +1226,24 @@ class SDK {
     debug('clone: Module version', version)
 
     try {
-      // 3 - after fetching module we still need to read the dat.json file
+      // 3 - after fetching module we still need to read the index.json file
       if (version !== 0) {
         moduleVersion = moduleDat.checkout(version)
       } else {
         moduleVersion = moduleDat
       }
-      debug('clone: Reading modules dat.json...')
+      debug('clone: Reading modules index.json...')
 
       const getFile = async () => {
         try {
-          return moduleVersion.readFile('dat.json', 'utf-8')
+          return moduleVersion.readFile('index.json', 'utf-8')
         } catch (_) {}
       }
 
       const content = await pRetry(getFile, {
         onFailedAttempt: error => {
           this._log(
-            `Failed attempt fetching dat.json. ${error.attemptNumber}/${
+            `Failed attempt fetching index.json. ${error.attemptNumber}/${
               error.retriesLeft
             }`
           )
@@ -1237,8 +1264,8 @@ class SDK {
     const modulePath = version ? `${mKeyString}+${version}` : `${mKeyString}`
     const folderPath = join(this.baseDir, modulePath)
     await ensureDir(folderPath)
-    await writeFile(join(folderPath, 'dat.json'), JSON.stringify(module))
-    const stat = await moduleDat.stat('dat.json')
+    await writeFile(join(folderPath, 'index.json'), JSON.stringify(module))
+    const stat = await moduleDat.stat('index.json')
 
     if (download) {
       dwldHandle = dat.downloadFiles(moduleVersion, folderPath)
@@ -1366,7 +1393,7 @@ class SDK {
     }
   }
 
-  async validatePublish (contentKey, profileKey) {
+  async validateRegister (contentKey, profileKey) {
     assert(
       typeof contentKey === 'string' || Buffer.isBuffer(contentKey),
       ValidationError,
@@ -1387,7 +1414,7 @@ class SDK {
     const { host: pKey, version: profileVersion } = parse(profileKey)
     if (!contentVersion) {
       this._log(
-        'publish: Content version is not found. Using latest version.',
+        'register: Content version is not found. Using latest version.',
         'warn'
       )
     }
@@ -1404,7 +1431,7 @@ class SDK {
 
     const { rawJSON: profile } = await this.clone(pKey, profileVersion, false)
 
-    // TODO(dk): consider add custom errors for publish and verification
+    // TODO(dk): consider add custom errors for registration and verification
     assert(
       content.type === 'content',
       ValidationError,
@@ -1459,17 +1486,17 @@ class SDK {
   }
 
   /**
-   * publish a content module to a profile
+   * register a content module to a profile
    *
    * @public
    * @async
-   * @link https://github.com/p2pcommons/specs/blob/master/module.md#registration
+   * @link https://github.com/p2pcommons/specs/blob/main/module.md#registration
    * @param {(String|Buffer)} contentKey - a dat url
    * @param {(String|Buffer)} profileKey - a dat url
    */
-  async publish (contentKey, profileKey) {
-    debug(`publish contentKey: ${contentKey}`)
-    debug(`publish profileKey: ${profileKey}`)
+  async register (contentKey, profileKey) {
+    debug(`register contentKey: ${contentKey}`)
+    debug(`register profileKey: ${profileKey}`)
 
     // profile.p2pcommons.contents.push(cKeyVersion)
     // update profile
@@ -1478,7 +1505,7 @@ class SDK {
       contents: [contentKey]
     })
 
-    this._log('publish: profile updated successfully')
+    this._log('register: profile updated successfully')
   }
 
   /**
@@ -1486,7 +1513,7 @@ class SDK {
    *
    * @public
    * @async
-   * @link https://github.com/p2pcommons/specs/blob/master/module.md#verification
+   * @link https://github.com/p2pcommons/specs/blob/main/module.md#verification
    * @param {String} datUrl - a versioned dat url
    * @returns {Boolean} - true if module is verified, false otherwise
    */
@@ -1517,12 +1544,12 @@ class SDK {
   }
 
   /**
-   * unpublish content from a user's profile
+   * deregister content from a user's profile
    *
    * @param {(String)} contentKey - contentKey should include the version
    * @param {(String|Buffer)} profileKey
    */
-  async unpublish (contentKey, profileKey) {
+  async deregister (contentKey, profileKey) {
     assert(
       typeof contentKey === 'string' || Buffer.isBuffer(contentKey),
       ValidationError,
@@ -1537,7 +1564,7 @@ class SDK {
       profileKey,
       'profileKey'
     )
-    debug('unpublish')
+    debug('deregister')
 
     await this.ready()
 
@@ -1788,11 +1815,11 @@ class SDK {
     }
     try {
       if (module.type === 'content') {
-        // unpublish from profiles
+        // deregister from profiles
         const profiles = await this.listProfiles()
 
         for (const { rawJSON: prof } of profiles) {
-          await this.unpublish(key, prof.url)
+          await this.deregister(key, prof.url)
         }
       }
       await this.localdb.del(keyString)
@@ -1834,26 +1861,39 @@ class SDK {
         mirror.destroy()
       }
     }
+
     if (db) {
       debug('closing db...')
-      if (this.localdb) await this.localdb.close()
-      if (this.seeddb) await this.seeddb.close()
-      if (this.db) await this.db.close()
+      try {
+        if (this.localdb) await this.localdb.close()
+        if (this.seeddb) await this.seeddb.close()
+        if (this.db) await this.db.close()
+      } catch (err) {
+        this._log(err.message, 'error')
+      }
       debug('db successfully closed')
     }
+
     if (swarm && this.networker) {
+      for (const drive of this.drives.values()) {
+        if (drive.closing && !drive.closed) {
+          await drive.close()
+        }
+      }
       debug('closing swarm...')
       try {
         await this.networker.close()
       } catch (err) {
         this._log(err.message, 'error')
       }
+
       debug('swarm successfully closed')
     }
+
     this.start = false
   }
 }
 
-SDK.errors = { ValidationError, InvalidKeyError, MissingParam }
+SDK.errors = { ValidationError, InvalidKeyError, MissingParam, EBUSYError }
 
 module.exports = SDK